j'ai des cartes youhouuuuuuuu

+++ src/node.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 Node::Node() : _x_min(0), _y_min(0), _x_max(0), _y_max(0), _triangles(0) {}
 
-Node::Node(double x_min, double y_min, double x_max, double y_max, vector<int> triangles) : _x_min(x_min), _y_min(y_min), _x_max(x_max), _y_max(y_max), _triangles(triangles) {
+Node::Node(double x_min, double y_min, double x_max, double y_max, vector<unsigned long> triangles) : _x_min(x_min), _y_min(y_min), _x_max(x_max), _y_max(y_max), _triangles(triangles) {
     _x_center = (x_min + x_max) / 2;
     _y_center = (y_min + y_max) / 2;
 }
